+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- DF-1222 Moving automatic cache flush for service to the ServiceManager
 
 ## [0.12.3] - 2017-09-19
 ### Fixed